Support for resetting directory-index
